Revert "Merge pull request #145 from NilFoundation/dev"
This reverts commit d1d45e04093ec067febd4444ae9f329edc361228, reversing
changes made to 9adb4cc83a8f62b308fbec0f919a2bd8b6fef5b7.
