+++ site/package.json
@@ -12,37 +12,28 @@
   },
   "dependencies": {
     "@hotjar/browser": "^1.0.9",
-    "@nilfoundation/ui-kit": "^2.2.11",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
     "axios": "^1.4.0",
-    "baseui": "^13.0.0",
     "classnames": "^2.3.2",
     "dayjs": "^1.11.9",
+    "fs": "^0.0.1-security",
     "gsap": "^3.12.1",
-    "html-to-text": "^9.0.5",
     "lodash.debounce": "^4.0.8",
     "lodash.flatten": "^4.4.0",
-    "lodash.uniq": "^4.5.0",
     "lottie-web": "^5.12.2",
     "next": "^13.4.7",
     "prop-types": "^15.8.1",
     "qs": "^6.11.2",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "sanitize-html": "^2.11.0",
-    "styletron-engine-atomic": "^1.6.2",
-    "styletron-react": "^6.1.1"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@biomejs/biome": "1.2.2",
-    "@types/html-to-text": "^9.0.4",
     "@types/lodash.debounce": "^4.0.7",
     "@types/lodash.flatten": "^4.4.7",
-    "@types/lodash.uniq": "^4.5.9",
     "@types/node": "^20.5.7",
     "@types/qs": "^6.9.7",
-    "@types/sanitize-html": "^2.9.5",
     "@types/showdown": "^2.0.1",
     "eslint": "8.16.0",
     "eslint-config-airbnb": "^19.0.4",
