+++ site/src/components/common/StickyContainer/StickyContainer.tsx
@@ -9,17 +9,14 @@ import s from './StickyContainer.module.scss'
 type StickyContainerProps = {
   children?: React.ReactNode
   className?: string
-  showWhiteRectangle?: boolean
 }
 
-const StickyContainer = forwardRef<HTMLDivElement, StickyContainerProps>(
-  ({ className, children, showWhiteRectangle = true }, ref) => (
-    <div ref={ref} className={cx(s.root, className)}>
-      {children}
-      {showWhiteRectangle && <WhiteRectangle className={s.whiteRectangle} />}
-    </div>
-  ),
-)
+const StickyContainer = forwardRef<HTMLDivElement, StickyContainerProps>(({ className, children }, ref) => (
+  <div ref={ref} className={cx(s.root, className)}>
+    {children}
+    <WhiteRectangle className={s.whiteRectangle} />
+  </div>
+))
 
 StickyContainer.displayName = 'StickyContainer'
