--- site/src/components/pages/OpenJobs/DottedSection/DottedSection.module.scss
@@ -1,114 +1,0 @@
-@import 'src/styles/typography';
-@import 'src/styles/helpers';
-@import 'src/styles/colors';
-@import 'src/styles/mixins';
-@import 'src/styles/paddings';
-
-.container {
-  background: url('/icons/dots.svg'), repeat;
-  display: grid;
-  height: size(620);
-  grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-
-  > div:first-child {
-    grid-column: 1 / 3;
-    grid-row: 1 / 3;
-    background-color: transparent;
-    position: relative;
-  }
-
-  > div:nth-child(3) {
-    grid-column: 1 / 3;
-    grid-row: 3 / 4;
-    background-color: $gray900;
-    position: relative;
-  }
-
-  @include mobile {
-    height: size(240);
-    margin: 0 -12px;
-    grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: repeat(4, 1fr);
-
-    & > div:first-child {
-      grid-column: 1 / 2;
-      grid-row: 1 / 5;
-    }
-
-    & > div:nth-child(3) {
-      grid-column: 2 / 3;
-      grid-row: 4 / 5;
-      background-color: transparent;
-    }
-  }
-}
-
-.whiteRectBottom {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-}
-
-.whiteRectTop {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-}
-
-.subgrid {
-  grid-column: 3 / 5;
-  grid-row: 2 / 4;
-  display: grid;
-  grid-template-columns: subgrid;
-  grid-template-rows: subgrid;
-  background-color: transparent;
-
-  > div:first-child {
-    grid-column: 1 / 2;
-    grid-row: 1 / 3;
-    background-color: transparent;
-    position: relative;
-  }
-
-  > div:last-child {
-    grid-column: 2 / 3;
-    grid-row: 1 / 3;
-    background-color: $gray900;
-    margin-top: size(280);
-    position: relative;
-  }
-}
-
-.titleWrapper {
-  position: relative;
-  grid-row: 1 / 2;
-  grid-column: 3 / 5;
-  padding: size(30) size(30) size(60);
-  background-color: $gray900;
-
-  @include mobile {
-    padding: 0 size(10) size($spacing32 + 14);
-    grid-row: 1 / 4;
-    grid-column: 2 / 3;
-  }
-}
-
-.title {
-  @include heading-03;
-
-  @include mobile {
-    @include main;
-  }
-}
-
-.link {
-  text-decoration: underline;
-  transition: opacity 0.3s ease;
-
-  @include hover {
-    opacity: 0.5;
-  }
-}
