--- site/src/components/pages/OpenJobs/Filter/Filter.module.scss
@@ -1,35 +1,0 @@
-@import 'src/styles/typography';
-@import 'src/styles/helpers';
-@import 'src/styles/colors';
-@import 'src/styles/mixins';
-@import 'src/styles/paddings';
-
-.filter {
-    display: grid;
-    gap: 32px;
-    grid-template-columns: 1fr 1fr 1fr;
-    grid-template-rows: 1fr 1fr;
-
-    @include mobile {
-        grid-template-rows: auto;
-        gap: size(16);
-        padding: size(24) 0;
-
-        & > * {
-            grid-column: span 3;
-        }
-    }
-}
-
-.remoteToggleContainer {
-    grid-column: 1 / 2;
-    grid-row: 2 / 3;
-
-    @include mobile {
-        grid-column: 1 / 4;
-    }
-}
-
-.titleTitleFilterContainer {
-    grid-column: 1 / 4;
-}
