--- site/src/components/pages/OpenJobs/Job/Job.module.scss
@@ -1,58 +1,0 @@
-@use '~styles/helpers' as *;
-@use '~styles/colors' as *;
-@use '~styles/typography' as *;
-@use '~styles/mixins' as *;
-@use '~styles/paddings' as *;
-
-.job {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 32px;
-    width: 100%;
-    height: 198px;
-
-    @include mobile {
-        grid-template-columns: 1fr;
-        grid-template-rows: auto;
-        gap: 16px;
-        height: auto;
-    }
-
-    &:not(:last-child) {
-        margin-bottom: 24px;
-    }
-
-    svg {
-        width: size(14);
-        height: size(14);
-    }
-}
-
-.type {
-    display: flex;
-    gap: 10px;
-}
-
-.title {
-    grid-column: 1 / 2;
-}
-
-.description {
-    grid-column: 2 / 3;
-    font-size: 18px;
-    font-weight: 400;
-    line-height: 24px;
-    color: $gray-300;
-    height: 48px;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    text-overflow: ellipsis;
-
-    @include mobile {
-        grid-column: 1 / 2;
-        -webkit-line-clamp: 4;
-        height: 96px;
-    }
-}
