--- site/src/components/pages/OpenJobs/index.ts
@@ -1,2 +1,0 @@
-export { default } from './OpenJobs'
-export { default as JobPage } from './JobPage/JobPage'
