+++ site/src/components/pages/OpenPositions/DottedSection/DottedSection.module.scss
@@ -1,0 +1,64 @@
+@import 'src/styles/typography';
+@import 'src/styles/helpers';
+@import 'src/styles/colors';
+@import 'src/styles/mixins';
+@import 'src/styles/paddings';
+
+.container {
+  background-color: $gray900;
+}
+
+.dottedCard {
+  position: relative;
+  padding-bottom: size(264);
+
+  @include mobile {
+    padding-bottom: size(80);
+  }
+}
+
+.line {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+}
+
+.bottomLine {
+  composes: line;
+}
+
+.wrapper {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.titleWrapper {
+  position: relative;
+  width: 50%;
+  padding: size(30) size(30) size(60);
+  margin-top: size(-30);
+  background-color: $gray900;
+
+  @include mobile {
+    padding: 0 size(10) size($spacing32 + 14);
+    margin-top: size(-22);
+  }
+}
+
+.title {
+  @include heading-03;
+
+  @include mobile {
+    @include main;
+  }
+}
+
+.link {
+  text-decoration: underline;
+  transition: opacity 0.3s ease;
+
+  @include hover {
+    opacity: 0.5;
+  }
+}
