+++ site/src/components/pages/OpenPositions/OpenPositions.module.scss
@@ -11,7 +11,6 @@
   @include mobile {
     display: flex;
     flex-direction: column;
-    padding: 0 12px;
   }
 }
 
@@ -34,7 +33,7 @@
   }
 
   @include mobile {
-    margin-top: size(24);
+    margin: size(10) size(20) size(18) size(16);
   }
 }
 
@@ -42,29 +41,4 @@
   transform: rotate(270deg);
   width: size(18);
   height: size(14);
-}
-
-.wrapper {
-  background-color: $gray900;
-  padding: 32px 0;
-
-  @include mobile {
-    padding: 24px 0 48px;
-  }
-}
-
-.department {
-  height: "64px";
-  margin-top: 64px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  @include mobile {
-    margin-top: 48px;
-    margin-bottom: 24px;
-    flex-direction: column;
-    align-items: flex-start;
-    height: auto;
-  }
 }
