+++ site/src/components/pages/OpenPositions/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OpenPositions'
