--- site/src/freshteam/index.ts
@@ -1,1 +1,0 @@
-export * from './api'
