+++ site/src/styles/globals.scss
@@ -40,5 +40,6 @@ a {
 }
 
 #freshteam-widget {
+  min-height: 100vh;
   background-color: $gray900 !important;
 }
